requirements.txt
asgiref==3.8.1
Django==5.1.3
python-decouple==3.8
sqlparse==0.5.2
tzdata==2024.2
